Add optional src dir
You can now configure a directory where the app looks for epub files, that is different from the output directory.
For example I can now use Nextcloud to save my epubs and have the file-system representation of the folder as my source directory thus my waiting time to index the directory and find the source epub is drastically reduced.

Signed-off-by: Simon K <Oct4v14n@googlemail.com>