Fix handling of Story- ViewModel and -DownloadService inconsistencies.
if a story was not entered via a ViewModel into the room database prior to calling the StoryDownloadService a NullPointerException would be thrown at the time the url is used to reference the non-existent entry in the room database.
Now a new entry will be added by the service itself.